LV522: Changes to map decortation, loot tables and renaming of colony airlocks (#4386)
# About the pull request

This PR

Updates some LV522 areas that were comparatively under-decorated
compared to the rest of the map
Heavily Nerfs the loose metal spawns on the map replacing them with
things like pouches and equipment
Renames alot of doors to have more accurate names

# Explain why it's good for the game

When I was making the map it was very clear that I didn't know what to
place in some areas so I just put down some loose metal, this has come
back to bite me in the ass with the FORECON tunnel hold and I wanna
knock that area down a peg without directly touching it as I still
believe it looks pretty cool so I've added some decor to make it take
longer to get into and build up and nerfed a lot of the potential metal
spawns

There's a new area between dorms and security I'll be calling it
Checkpoint Charlie in this PR but it goes unnamed in the map file.
Checkpoint Charlie is the last stand location for FORECON's GL spec and
thus has some explosives FORECON can pick up and use such as an M79, 3
HEDPs, an M15, and a demo charge for getting into the crashed dropship.

The Landed FORECON ship has been made slightly easier to get into, now
only requiring a breaching charge to open instead of pushing a button
across the map, hopefully this will make FORECON consider getting whats
inside and the crashed ship has been slightly buffed with an
Autoinjector pouch and a compact defib for FORECOn to use.

The doors were renamed in an attempt to help people navigate and is
something that should have been in from the start


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: SpartanBobby, GSDPathe
maptweak: Changes to LV522 loot table surrounding metal spawns
maptweak: Redectorates areas around LV522 that were underwhelming before
/:cl: