balance and fix to reduce the potentiel of requisiton players to get Carpal tunnel syndrome (#4303)
# About the pull request
i made req have double what a single squad prep as in term of standard
gun and ammo.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
reduce the number of clicks and Carpal tunnel syndrome risk
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: made it so that req as double what a standard vendor as in ammo
and guns.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>