Removes some PMC gear (#4420)
# About the pull request

These should not be in areas that can be rushed by the same three people
every round.

# Explain why it's good for the game

Identifying PMC gear (specifically gloves, mask) is a huge pain as a
xeno and the difference it makes in survivability is huge. Really some
of these items just need toned down in general but these in specific
gotta go.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Removed some PMC gear on maps
/:cl: