Mateba Special (#4441)
# About the pull request

This PR adds an optional version of the mateba: the Mateba Special.

Based on an older version of the game's mateba.

Check the icondiffbot in checks for the sprites.

# Explain why it's good for the game

Neato


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow, DrLance (sprites)
add: Added the Mateba Special as a choice for COs
/:cl: