Removes eggs off hive weeds (#4459)
# About the pull request

This PR removes eggs off hive weeds.

# Explain why it's good for the game

Too strong, too many awkward layering issues to be abused.

I wanted to see if it'd create some sort of niche for egg sac carrier as
carrier in general has been kind of left behind on the power curve
especially since shaman removal. Sadly, it just wasn't meant to be.

As I said in #4015: "We'll revisit this in a couple weeks." Here we are.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: Removed eggs off hive weeds for egg sac carrier
/:cl: