Fix afk hugger conversion (#4487)
# About the pull request

This PR is a followup to #3886 fixing dead huggers still converting to
normal huggers when client-less. The assumption was that life wouldn't
keep ticking when dead, but thats not the case.

# Explain why it's good for the game

Fixes #4486 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

# Changelog
:cl: Drathek
fix: Fix dead huggers converting to non-player huggers
/:cl: