Fixes akimbo bug (#4503)
# About the pull request

Fixes accidental infinite firing loop when firing akimbo.

In testing I was using pump shotguns and akimbo firing is rather rare in
the first place so didn't pick up on it.

# Explain why it's good for the game

Firing infinitely to kill xenos instantly is bad.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Fixed a bug with akimbo firing
/:cl: