Yautja Rifle buff (#4222)
# About the pull request
While the plasma rifle did just get an update to it's capabilities, it
still isn't hitting xenos for very much considering their health pools
are much larger than a human's. This adds an extra 75% damage to xenos,
and also applies hive interference to make it more consequential.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes the rifle a little more useful against xenos
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Adds an extra 75% damage and hive interference to plasma rifle
vs xenos.
/:cl: