Natural Trijent Dam Tunnels (#4458)
# About the pull request

Adds 9 naturally spawning tunnels to Trijent Dam.

# Explain why it's good for the game

Trijent Dam is a map which has a complete lack of naturally spawning
tunnels unlike practically every other map, this is addition makes it
fall in line with all of the other maps having naturally spawning
tunnels and makes it less of a chore to set up a proper tunnel system
during normal gameplay.

# Testing Photographs and Procedure

Ran it locally, maps are in intended places and function as a normal
tunnel would.


![image](https://github.com/cmss13-devs/cmss13/assets/31109792/a0a5d938-9269-4fbf-a842-d2574ac19002)

The tunnel locations.

![Trijent Tunnel Map V2
1](https://github.com/cmss13-devs/cmss13/assets/31109792/3eb9fbf2-09fd-4c3b-949e-7d52f93b60b5)

# Changelog

:cl:
mapadd: added 9 new tunnels to trijent dam
/:cl: