Removes alpha egg on LV522 (#4508)
# About the pull request

Removes alpha egg on LV522.

# Explain why it's good for the game

When it gets to the point of metarushing the egg to fuck with rounds
it's time to go. Secrets can only be secrets when they're not being
passed around to abuse. RIP a real one.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
del: Removed alpha egg on LV522
/:cl: