Auto observer fixes, preferences, and QoL (#4463)
# About the pull request

This PR creates wrappers for screen additions and removals to allow for
a creation of a signal observers can then register to so the observed
screen changes are all changed on any observer's screen as well.

Let me be very clear: this needs a lot of testing and I did this kind of
ad hoc just to see if it would work. Brief testing seems to show it
working but I expect there are oversights and problems. More than likely
I am recreating the wheel in some way and doing it the dumb way.

# Explain why it's good for the game

Making the HUD observer more accurate is cool.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
add: Added a save-able preference for auto-observe
fix: Fixed numerous problems with HUD observer mode
code: Created a wrapper for client screen changes
code: Made check contents less horrendous
/:cl: