Fixes combistick folding (#4526)
# About the pull request
Lets combisticks be folded without using the verb menu by splitting it
into a new verb/keybind

# Explain why it's good for the game
The verb menu is terrible, and it was fully intended for this to be
macro/keybind-able

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

works

</details>


# Changelog
:cl:
add: You can now fold a combi-stick using the new "collapse combi-stick"
verb and/or keybind. Defaults to the space bar.
/:cl: