Fixes taking control of crit xenos (#4536)
# About the pull request

Alright so I did not consider that *all* xenos that did not have a
client and could be taken over would be considered unconscious.

# Explain why it's good for the game

Uuuh you should be able to join as a xeno...


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: Fixed taking control of crit xenos
/:cl: