Tool Pouch Allowed Items Change (#4554)
# About the pull request

The tool pouch can now hold the same items that the toolbelt is allowed
to hold. It is still restricted to four item slots.

# Explain why it's good for the game

Consistency in item management. It felt odd the tool pouch was excluded
from holding items the toolbelt can hold.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Tool Pouch can now hold the same types of items that a toolbelt
can hold. It is still restricted to 4 inventory slots.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>