Allows Xeno Hive To Ally With CMB (#4562)
# About the pull request

See title, now allows a xenomorph hive to ally with Colonial Marshals. 

# Explain why it's good for the game

Oversight of a new faction not being able to be allied with, can create
scenarios where friendly marshals are excluded from an alliance and can
be attacked by other xenomorphs, even if the Queen does not want this to
occur.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Xenomorph hives can now form an alliance with Colonial Marshals. 
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>