"Fixes" missing window frame icon states (#4534)
# About the pull request

This PR "fixes" missing window frame icon states by just checking if
they exist and if not then defaulting to base.

The vast majority of these are missing states 5, 6, 7, 9, 10, 11, 13,
14, 15 but some have them and others do not.

It's essentially a bandaid but until we get someone who wants to sprite
100ish icons 🤷

closes https://github.com/cmss13-devs/cmss13/issues/2531

# Explain why it's good for the game

Dealing with invisible window frames is super frustrating for everyone.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
fix: "Fixed" missing window frame icon states
/:cl: