Xeno structure cheese reduction (#4550)
# About the pull request

If a mob is on top of a reflective wall the wall does not function.

If you are not directly clicking on a xeno special structure
(cluster/core/etc) it will not stop your bullets.

# Explain why it's good for the game

Seeing an uptick in cheesy nonsense we don't really want to encourage.
This should mitigate it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
balance: If a mob is on top of a reflective wall the wall no longer
functions
balance: Bullets will not longer get eaten by xeno special structures
unless you click directly on them
/:cl: