UPP surv redo and removing ambiguous text (#4576)
# About the pull request

The previous UPP surv changes got undone in a soft conflict while moving
things around so they're back.

The removed text creates a feeling of *possibly* the USCM are hostile
but this is never the case and UPP are explicitly not hostile as
survivors.

# Explain why it's good for the game

Not losing PRs good.

Confusing text bad.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Morrow
fix: Re-added previous UPP changes that were lost in soft conflict
spellcheck: Removed some ambiguous text about loyalties for UPP
survivors
/:cl: