Minor QOL Buff To Handheld Crew Monitor (#4545)
# About the pull request

Adds the job title of a person to the crew list on the handheld crew
monitor. Also expands the default screen size of the radar menu so the
user can see the entire name and job title of the crew.

# Explain why it's good for the game

Makes the monitor slightly easier to use. Helps differentiate who
exactly you are looking at.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2023-09-29 14 31
45](https://github.com/cmss13-devs/cmss13/assets/6595389/6bc15031-50e9-475c-9d00-22e13979fd38)


</details>


# Changelog
:cl:
ui: Added personnel job titles to the handheld crew monitor, to make it
easier to tell find out exactly what role's you are looking for. Also
slightly expands the default monitor screen size to accommodate the
entire personnel text on screen.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>