Phone Do not Disturb (#4427)
# About the pull request
Allows phones to be put on Do Not Disturb, preventing them from being
called. This is not possible to toggle on certain phones such as
Overwatch.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Sometimes phones get spammed for trolls, other times the relevant people
(like CL) are not in their office. This makes it easy to stop wasted
calls on empty rooms or busy people.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a Do Not Disturb feature for MOST phones. (Overwatch and a
couple others excluded).
/:cl: