tgui research desktop fix (#4592)
# About the pull request
3 out of 4

Have you ever noticed that when you simulate something new, lets say you
called it "gamer juice" and instead, in computer lab it always has "for"
instead of the actual name that you created? this is because "Analysis
of X" Has two words before the name of the chemical while "Simulation
results for X" has 3

TGUI splits it the same way despite having different number of words,
thus, "Simulation results **for** X" - this is where For comes from.

I had two ways to do it, either remove third word, or try to fix it in
tgui, this however, still doesnt include the spaces you can put in the
name.

(tgui is weiiirdd, if I did something funny which I dont get, sorry)
# Explain why it's good for the game
Always annoying to look for correct chemical you did in like ~8 "for"


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: kiVts
fix: Research computer no longer shows "for" for every simulation
result.
/:cl: