Resprites Tactical shotgun stock +  stock now comes standard on all tactical shotguns (yes you can remove it) (#4444)
# About the pull request

See title

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

It's there, sitting, unused for years, yes it kinda sucks but the sprite
is good and it makes the shotgun look cooler so we might as well let it
be used

https://i.imgur.com/erbgUaK.png

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: SpartanBobby, esselnek
add: tactical shotguns now come standard with their stock. Yes you can
remove it if you dont want it
add: new tactical shotgun stock sprite done by esselnek
/:cl: