Slightly optimizes client performance (#4610)
# About the pull request
Ports https://github.com/DaedalusDock/daedalusdock/pull/638, credit to
Kapu for the PR

# Explain why it's good for the game
By having the `icon` be `null` (instead of `icon_state` being `""`),
BYOND processes one less appearance (that of the invisible area) per
turf. That being said, appearance-to-icon is cached, but this is free
performance.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

It doesn't fuck with weather or anything

</details>