Small changes to light floors (#4599)
# About the pull request

Adds subtypes to light floors and examine text
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes it easier for mapping

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

I didn't test it

</details>


# Changelog
:cl:
add: Light floor can be examined to tell if it is on, off, or broken.
code: Light floor subtypes for easier mapping
/:cl: