Colonial Marshals in Distress! CMB Handheld Distress Beacon (#4540)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Adds CMB Distress Beacon which is given to CMB Marshal. Which represent
a set of HRP roles.
This is pressed and turns on, sending a message to admins.
Admins can reply, deny, or choose to send one or both ERTs available to
them.

First ERT is the standard combat response, Anchorpoint Station Colonial
Marine QRF. 6 Marines ready to rescue CMB.
Second ERT is a very poor combat response. CMB Patrol Team - Marshals in
Distress. Up to 5 CMB people of various types which are not prepared for
combat. This may be better for a smaller scale response.

Denying the Distress Beacon gives a little message which basically says
"NO QRF is available, sorry. I'll relay your request. Good luck."

Replying will give a human voice through the radio headset which PMs the
player but ICly.

Down the line I **MIGHT** add a CMB "SWAT" team as featured which would
still be weak but slightly stronger than the patrol team(ie, they'd have
lights and possibly the integrated armor and 37-12 flamers) but right
now I don't see the need.

Special thanks to Morrow

# Explain why it's good for the game
There is a significant amount of mechanical issue when it comes to fax
interactions an admins especially towards the end of the game. Admins
may not have time to write faxes and the players might not either. In
addition to call things it can be convoluted or uncertain at times.

This clears things up a lot. A player comes to a conclusion that they
may need assistance. An admin can then teleport to the player, and
choose a variety of responses such as a radio reply, denial, or sending
in various types of reinforcements to the CMB team signalling distress.
It is mechanically more fluid and also immersive.

I'm also hoping that this will allow the Anchorpoint Marine QRF to
assist the Marshals as intended in their 2-part ERT.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

https://media.discordapp.net/attachments/975247220960014386/1157041837941473442/image.png?ex=65172aae&is=6515d92e&hm=a5fb779e0bd74238d9334b67fd3b6ea1961deef2dd63d025df832a3b3e40e01f&=

https://media.discordapp.net/attachments/975247220960014386/1157045204034015324/image.png?ex=65172dd1&is=6515dc51&hm=754a2e613319534afc82b099b217ae98227deb49f3b9c891adffac3cc810d40e&=

https://cdn.discordapp.com/attachments/987279688282021908/1157041436538196048/image.png?ex=65172a4f&is=6515d8cf&hm=44758548407d0d9394a2437764ad9c65f5d685528e0c292dda8ab9ab3ef52b98&

https://cdn.discordapp.com/attachments/987279688282021908/1157038354488107008/image.png?ex=65172770&is=6515d5f0&hm=7a7195bb3a2d543ffad2db151a175f2d897b9edca35f5fe98a9e19cd0ccba1be&

</details>
slightly outdated screenshots in terms of desc/lore/writing, but
mechanically the same - taken during good tests.

# Changelog
:cl:
add: Adds a handheld distress beacon for the Colonial Marshal. They can
use this to signal distress which in turn allows reinforcements in the
form of Anchorpoint Marine QRF or nearby CMB teams. Admins do the final
check.
add: Adds a CMB Patrol Team which responds to "Marshals in Distress"
call.
admin: headset admin response no longer specifies USCM origin(as this is
used for USCM, WY, and now CMB)
spellcheck: fixes a CMB typo in response message. Also a miniscule
change to HG 37-12 desc.
spellcheck: fixes 2 misc typos in CMB Inspections
/:cl: