changed flyby tgui (#4600)
# About the pull request

Slight change to the TGUI for dropship flight controls. I've had
feedback that it is a bit unintuitive. This change should resolve novice
POs accidentally launching flyby in a panic.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Better UIs make everyone happier.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.


![image](https://github.com/cmss13-devs/cmss13/assets/101871009/87c93141-1ba8-4cf0-b8d0-3bf54b97844e)


</details>


# Changelog
:cl:
ui: tweaked flyby controls to make them in line with other destinations
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>