Photo Code Refactoring (#4636)
# About the pull request

This PR is a partial revival of #4584 in addition to just a general
refactor of the code it was affecting. Now after every blend or
getFlatIcon call there is a CHECK_TICK. Various loops use the `as
anything` usage as possible to help with type checking. Additionally,
the descriptions for mobs is a little more fluid, and now there is
actually a description for a photo with no mobs in it.

# Explain why it's good for the game

This should hopefully mitigate issues where photos cause the server to
stutter, but hard to know for sure without testing on live. Also the
description was annoyingly rigid.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/4d3d5dc4-bd1e-4a63-a188-b0a21076ee99)
(There was an additional space in this photo's description that was
since removed)

</details>

# Changelog
:cl: Drathek
refactor: Refactored camera code to be less blocking, use typechecks
less often, and provide somewhat more fluid descriptions to photos.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>