Fixes nuke explosion runtiming if a living mob was in a nullspaced container (#4659)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

As per title.

I have no idea what causes this condition to begin with, live inspection
didn't yield any results.

# Explain why it's good for the game
This might sound like an improbable case, and it is, but it actually
happened last week. This caused the nuke explosion to stop and left
everyone confused for the next 45 minutes the round lasted to end
normally.

# Testing Photographs and Procedure
Untested.


# Changelog
:cl:
fix: Fix nuke explosion getting canceled in some very rare cases.
/:cl: