Fixes Tape Recorder failing to record when no language is given (#4663)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

The very same proc has a check on the language existing two line further
up, because it's a potential use cases -- but fails to check it again
and runtimes further below.


# Explain why it's good for the game
Bugfix

# Testing Photographs and Procedure
Untested


# Changelog
:cl:
fix: Fixed Tape Recorder failing to record when no language information
is present.
/:cl: