Fixes Disconnected Observers runtiming HUD (#4667)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

A simple runtime job... Login() to an observed mob with a disconnected
observer will crash the inventory HUD proc because it has no client.

# Explain why it's good for the game
Less runtimes

# Testing Photographs and Procedure
Tested with dual client and breakpoints


# Changelog

Internal change only