Gives ERTs a variant of the regular medkit that isn't locked to skills in packs. (#4670)
# About the pull request

A bunch of ERTs are designed with having their meds in medkits, which
was obviously not noticed or cared about when making !4443 so I made a
variant of the regular kit that has no lock, only ERT roles(and 2 HG
roles on WO) have it so its not a real balance concern.

# Explain why it's good for the game

Storage nesting is not an issue with normal preset loadouts like ERTs(or
at all :D) as they are designed around and only have access to specific
loadouts.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: ERTs now have a variant of the standard medkit that has no
storage lock.
fix: Lots of ERTs now correctly have a medkit spawning in their gear
/:cl: