Fix join as xeno observer alert text url (#4681)
# About the pull request

This PR simply sets an argument for the join as xeno href that the hive
surge observer action provides in chat so the href topic will actually
respond to it if clicked.

# Explain why it's good for the game

Links to things should actually work.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixed the hive surge join as xeno observer alert text link not
actually doing anything.
/:cl: