Nerfs the black market (#4674)
# About the pull request

- Removed several guns from the black market
- Removed extended m4ra mags from black market
- Removed monkeycubes from the "food" crate
- The black market now requires a specialized tool (sprite by Wei) to
hack. Can be found in the scanner object spawner 2.5% of the time. 9
spawners exist on the Almayer, meaning one will normally spawn every 2-3
rounds.
- Removed lunge mines from black market

# Explain why it's good for the game
The black market is not in a good place currently.

- It is completely trivial to access the black market, owing to the ease
of hacking and the CL's ability to print money. This removes any chance
of it being a special thing, instead turning it into an affair that
occurs almost every round.
- The CMB don't actually fix anything, they just encourage CTs to fuck
around with the black market
- Many of the things in the black market are just better than marine
gear. Removing the really good weapons is the start, I will have to see
if anything more is necessary.
- Monkey cubes allow for effectively infinite greenoes. The black market
shouldn't be significantly impacting a round, and this is directly
counter to that.
- Lunge mines are dumb as fuck, instantly gib humans (yes even the
broken ones do that) and encourage LRP behavior.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/41448081/a5f28f12-f660-4d0d-944d-ecfba40879fc)

</details>


# Changelog
:cl: Zonespace, Wei (sprites)
balance: Removed NSG, Type71, tactical Mar-40, and Mar-50 from the black
market
balance: Removed Ext. M4ra mags from the black market
balance: Removed monkeycubes from the black market
balance: Removed lunge mines from the black market
balance: You now need a modified security access tuner to access the
black market, which can be found ship- and ground-side where scanners
and radios can spawn.
/:cl: