Allows MT, CT, and Nurse to take the "Bad Leg" trait. (#4704)
# About the pull request

This does as the title says, allowing Maintenance Techs, Cargo Techs,
and Nurses to take the Bad Leg trait.

# Explain why it's good for the game

The reason that these roles *specifically* couldn't take this trait was
due to them being "too rookie" which I think is a bit silly to say. Even
if we take all three of these to be roles only new players play (they
aren't, definitely not MT or CT) - the player still intentionally had to
go in and take the trait and therefore is assumed to accept the traits
mechanics. If they have an issue with the trait or handling it because
they are new then that is an issue that can be resolved by them not
taking detrimental traits.

Forgive me if the cl tagging is wrong, this is not my codebase this just
annoyed me enough to PR it.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
qol: MT, CT, and Nurses can now take the bad leg trait.
/:cl: